requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-multipart==0.0.6

# Data processing
numpy==1.26.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Caching & Rate Limiting
redis==4.6.0  # Downgraded for compatibility with fastapi-cache2
fastapi-cache2[redis]==0.2.1
fastapi-limiter==0.1.5

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
mypy==1.7.0
types-python-dateutil==2.8.19.14
types-requests==2.31.0.20240106

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy-extensions==1.0.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.5